+++ sources/Google.Solutions.IapDesktop.Application.Test/Windows/Authorization/TestAccessInfoViewModel.cs
@@ -22,7 +22,7 @@
 using Google.Solutions.Apis.Auth;
 using Google.Solutions.Apis.Client;
 using Google.Solutions.IapDesktop.Application.Host.Adapters;
-using Google.Solutions.IapDesktop.Application.Windows.Authorization;
+using Google.Solutions.IapDesktop.Application.Windows.Auth;
 using Google.Solutions.Testing.Application.Test;
 using Moq;
 using NUnit.Framework;
