+++ sources/Google.Solutions.IapDesktop.Application/Windows/Auth/AccessInfoFlyoutView.resx
