+++ sources/Google.Solutions.IapDesktop.Application/Windows/Auth/AuthorizeView.resx
