+++ sources/Google.Solutions.IapDesktop.Application/Windows/Auth/AuthorizeViewModel.cs
@@ -40,7 +40,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Google.Solutions.IapDesktop.Application.Windows.Authorization
+namespace Google.Solutions.IapDesktop.Application.Windows.Auth
 {
     public class AuthorizeViewModel : ViewModelBase
     {
