+++ sources/Google.Solutions.IapDesktop.Application/Windows/Auth/NewProfileView.resx
