+++ sources/Google.Solutions.IapDesktop.Application/Windows/Auth/OAuthScopeNotGrantedView.resx
